Se modificaron los estilos con clases bootstrap en los componentes de book, se corrigieron detalles tanto en el formulario de edicion como el de creacion para los select option. Se modificaron los mensajes de respuesta del backend

+++ src/app/components/book-detail/book-detail.component.ts
@@ -24,13 +24,15 @@ export class BookDetailComponent implements OnInit {
   }
 
   deleteBook(id) {
-    this.http.delete('/api/book/'+id)
+    if(window.confirm("Are you sure to delete?")) {
+      this.http.delete('/api/book/'+id)
       .subscribe(res => {
           this.router.navigate(['/books']);
         }, (err) => {
           console.log(err);
         }
       );
+    }
   }
 
 }
